Update against master
This document is still ropey as heck, but tries to do away with the
misinformation present in the original document. It begins to loosely
output the scene referred to display referred pipeline. Someone with
more experience specifically with Godot should flesh out how textures
are handled specifically via the UI, and what toggles to set.

In the end, Godot desperately needs to begin pixel managing the pipeline, at
which point this document can be better fleshed out.

Still, a baby step in a more appropriate direction.